Change fileBmi5

+++ src/Main.java
@@ -2,8 +2,8 @@ public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
         double height = 1.87;
-        int weight = 98;
-        double index = (int) (service.calculate(height, weight));
+        double weight = 98;
+        double index = service.calculate(height, (int) weight);
         System.out.println("ИМТ(Индекс Массы Тела) = " + Math.round(index));
